refactor(tasks): migrate summarize module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
