+++ src/pyrovelocity/tasks/__init__.py
@@ -1,6 +1,7 @@
 import pyrovelocity.tasks.data
 import pyrovelocity.tasks.postprocess
 import pyrovelocity.tasks.preprocess
+import pyrovelocity.tasks.summarize
 import pyrovelocity.tasks.train
 
 
@@ -8,5 +9,6 @@ __all__ = [
     "data",
     "postprocess",
     "preprocess",
+    "summarize",
     "train",
 ]
